requirements.txt
asgiref==3.3.1
astroid==2.5.1
attrs==20.3.0
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
djangorestframework==3.12.2
docopt==0.6.2
et-xmlfile==1.0.1
future==0.18.2
housecanary==0.7.1
idna==2.10
iniconfig==1.1.1
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
oauthlib==3.1.0
openpyxl==3.0.7
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pylint==2.7.2
pyparsing==2.4.7
pytest==6.2.2
python-slugify==4.0.1
pytz==2021.1
requests==2.25.1
requests-mock==1.8.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
